requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
-e git+https://github.com/DankMemes4President/rt-two-way-sync.git@49f00a14111acdb2c84165e890da19fd05f9a7bf#egg=event_listener&subdirectory=event_listener
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
stripe==8.8.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
